Working on data preprocessing
We have changed the names of the train files for example train_1, to
make it more readable. We have unpickled the train and test data for
CIFAR-10, and combined the 5 training data sets. We have also started
the check data function, code is all working up to checking the image
dimensions.